remove circular dependency on bit_length and orderlen
since bit_length and orderlen are utility methods, they should not
be defined in main modules (curves) or in modules for specific features
(rfc6979)

move them to the util module, keep them importable from old places though
